Ytupload/requirements.txt
flask>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
moviepy>=1.0.3
opencv-python>=4.6.0
numpy>=1.21.0
python-dotenv>=0.19.0
Werkzeug>=2.0.0
